monitorinterface/serializers.py
from rest_framework import serializers
from monitorinterface.models import Host, Metric, Measurement


class HostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Host
        fields = ('id', 'ip')


class MetricSerializer(serializers.ModelSerializer):
    class Meta:
        model = Metric
        fields = ('id', 'type', 'period_seconds')


class MeasurementSerializer(serializers.ModelSerializer):
    class Meta:
        model = Measurement
        fields = ('id', 'value', 'timestamp')
