requirements.txt
dj-database-url>=0.4.1
Django>=2.0.3
gunicorn>=19.6.0
psycopg2>=2.7.3.2
whitenoise>=3.2
djangorestframework